# agrold_webapp/web/swagger/agrold-2.json
{
    "swagger": "2.0",
    "info": {
        "description": "These are the documentations of the services the 1.0 version of the RESTful-like API of AgroLD reachable at http://volvestre.cirad.fr:8080/aldp/api/1.0",
        "version": "1.0",
        "title": "AgroLD API 1.0 Interactive Documentations",
        "termsOfService": "http://www.agrold.org",
        "contact": {
            "email": "tagnyngompe@gmail.com"
        },
        "license": {
            "name": "IBC Montpellier",
            "url": "http://www.ibc-montpellier.fr/wp/wp5"
        }
    },
    "host": "volvestre.cirad.fr:8080",
    "basePath": "/aldp/api/1.0",
...
    "schemes": [
        "http"
    ],
    "paths": {
        "/id/byOntoTerm{_format}": {
            "post": {
                "tags": [
                    "ontologies"
                ],
                "summary": "Returns all the IDs corresponding to an ontological term",
                "description": "&lt;n&gt; PREFIX rdfs:< http://www.w3.org/2000/01/rdf-schema# ><br>SELECT ?id <br>WHERE {<br>   ?subject rdfs:label \"homoaconitate hydratase activity\"^^xsd:string .<br>  BIND(REPLACE(str(?subject), '^.*(#|/)', \"\") AS ?localname)<br>  BIND(REPLACE(?localname, \"_\", \":\") as ?id).<br>} limit 1",
                "operationId": "getIdByOntoTerm",
                "produces": ["text/html", "application/sparql-results+json", "application/ld+json", "application/sparql-results+xml", "text/tab-separated-values", "text/csv", "rdf/xml", "text/turtle"],
                "parameters": [
                    {
                        "name": "ontoTerm",
                        "in": "query",
                        "description": "The ontological term (e.g. homoaconitate hydratase activity)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "number of the page of the result set to display (from 1 and greater than 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_pageSize",
                        "in": "query",
                        "description": "size of a page of the result set (> 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_format",
                        "in": "path",
                        "description": "format in which the results set will be return",
                        "required": true,
                        "type": "string",
                        "enum": [
                            ".json",
                            ".sparql-json",
                            ".xml",
                            ".tsv",
                            ".csv",
                            ".rdf",
                            ".html",
                            ".ttl"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/ontoTerm/byId{_format}": {
            "post": {
                "tags": [
                    "ontologies"
                ],
                "summary": "Returns the name of an ontological element corresponding to its given ID",
                "description": "PREFIX rdfs:< http://www.w3.org/2000/01/rdf-schema# ><br>SELECT ?id <br>WHERE {<br>   ?subject rdfs:label \"homoaconitate hydratase activity\"^^xsd:string .<br>  BIND(REPLACE(str(?subject), '^.*(#|/)', \"\") AS ?localname)<br>  BIND(REPLACE(?localname, \"_\", \":\") as ?id).<br>} limit 1",
                "operationId": "getOntoTermById",
                "produces": ["text/html", "application/sparql-results+json", "application/ld+json", "application/sparql-results+xml", "text/tab-separated-values", "text/csv", "rdf/xml", "text/turtle"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "The id (e.g. GO:0003824)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "number of the page of the result set to display (from 1 and greater than 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_pageSize",
                        "in": "query",
                        "description": "size of a page of the result set (> 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_format",
                        "in": "path",
                        "description": "format in which the results set will be return",
                        "required": true,
                        "type": "string",
                        "enum": [
                            ".json",
                            ".sparql-json",
                            ".xml",
                            ".tsv",
                            ".csv",
                            ".rdf",
                            ".html",
                            ".ttl"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/ancestors/byId{_format}": {
            "post": {
                "tags": [
                    "ontologies"
                ],
                "summary": "Returns all the IDs corresponding to an ontological term",
                "description": "<br> SELECT ?ancestorId<br>WHERE<br>{<br>#level 4<br>{<br>SELECT ?ancestor1<br>WHERE<br>{<br>?subject rdfs:subClassOf ?ancestor1.<br>FILTER REGEX(STR(?subject), CONCAT(REPLACE(\"<b>id</b>\", \":\", \"_\"), \"$\"))<br>} limit 1}<br>?ancestor1 rdfs:subClassOf ?ancestor2.<br>?ancestor2 rdfs:subClassOf ?ancestor3.<br>?ancestor3 rdfs:subClassOf ?ancestor4.<br>BIND(REPLACE(str(?ancestor4), '^.*(#|/)', \"\") AS ?ancestorLocalname)<br>BIND(REPLACE(?ancestorLocalname, \"_\", \":\") as ?ancestorId)<br>}",
                "operationId": "getAncestorById",
                "produces": ["text/html", "application/sparql-results+json", "application/ld+json", "application/sparql-results+xml", "text/tab-separated-values", "text/csv", "rdf/xml", "text/turtle"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "The ID of an ontological resource (e.g. GO:0004409 or G0:0016836)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "description": "The level of the ancestor (e.g. the level of the direct parent is 1)",
                        "required": true,
                        "type": "integer",
                        "minimum": 1
                    },
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "number of the page of the result set to display (from 1 and greater than 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_pageSize",
                        "in": "query",
                        "description": "size of a page of the result set (> 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_format",
                        "in": "path",
                        "description": "format in which the results set will be return",
                        "required": true,
                        "type": "string",
                        "enum": [
                            ".json",
                            ".sparql-json",
                            ".xml",
                            ".tsv",
                            ".csv",
                            ".rdf",
                            ".html",
                            ".ttl"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/parents/byId{_format}": {
            "post": {
                "tags": [
                    "ontologies"
                ],
                "summary": "Returns the parents of an ontological element given its id",
                "description": "<br> SELECT ?ancestorId<br>WHERE<br>{<br>{<br>SELECT ?ancestor1<br>WHERE<br>{<br>?subject rdfs:subClassOf ?ancestor1.<br>FILTER REGEX(STR(?subject), CONCAT(REPLACE(\"<b>id</b>\", \":\", \"_\"), \"$\"))<br>} limit 1}<br>?ancestor1 a owl:Class .<br>BIND(REPLACE(str(?ancestor1), '^.*(#|/)', \"\") AS ?ancestorLocalname)<br>BIND(REPLACE(?ancestorLocalname, \"_\", \":\") as ?ancestorId)<br>}",
                "operationId": "getParentById",
                "produces": [],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "The ID of an ontological resource (e.g. GO:0004409)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "number of the page of the result set to display (from 1 and greater than 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_pageSize",
                        "in": "query",
                        "description": "size of a page of the result set (> 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_format",
                        "in": "path",
                        "description": "format in which the results set will be return",
                        "required": true,
                        "type": "string",
                        "enum": [
                            ".json",
                            ".sparql-json",
                            ".xml",
                            ".tsv",
                            ".csv",
                            ".rdf",
                            ".html",
                            ".ttl"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/descendants/byId{_format}": {
            "post": {
                "tags": [
                    "ontologies"
                ],
                "summary": "Returns the descendents of an ontological element given its id",
                "description": "<br> SELECT ?ancestorId<br>WHERE<br>{<br>{<br>SELECT ?ancestor1<br>WHERE<br>{<br>?subject rdfs:subClassOf ?ancestor1.<br>FILTER REGEX(STR(?subject), CONCAT(REPLACE(\"<b>id</b>\", \":\", \"_\"), \"$\"))<br>} limit 1}<br>?ancestor1 a owl:Class .<br>BIND(REPLACE(str(?ancestor1), '^.*(#|/)', \"\") AS ?ancestorLocalname)<br>BIND(REPLACE(?ancestorLocalname, \"_\", \":\") as ?ancestorId)<br>}",
                "operationId": "getDescendentById",
                "produces": ["text/html", "application/sparql-results+json", "application/ld+json", "application/sparql-results+xml", "text/tab-separated-values", "text/csv", "rdf/xml", "text/turtle"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "The ID of an ontological resource (e.g. GO:0003824)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "description": "The level of the descendent (e.g. the level of the direct children is 1)",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "number of the page of the result set to display (from 1 and greater than 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_pageSize",
                        "in": "query",
                        "description": "size of a page of the result set (> 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_format",
                        "in": "path",
                        "description": "format in which the results set will be return",
                        "required": true,
                        "type": "string",
                        "enum": [
                            ".json",
                            ".sparql-json",
                            ".xml",
                            ".tsv",
                            ".csv",
                            ".rdf",
                            ".html",
                            ".ttl"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/children/byId{_format}": {
            "post": {
                "tags": [
                    "ontologies"
                ],
                "summary": "Returns the children of an ontological element given its id",
                "description": "<br> SELECT ?ancestorId<br>WHERE<br>{<br>{<br>SELECT ?ancestor1<br>WHERE<br>{<br>?subject rdfs:subClassOf ?ancestor1.<br>FILTER REGEX(STR(?subject), CONCAT(REPLACE(\"<b>id</b>\", \":\", \"_\"), \"$\"))<br>} limit 1}<br>?ancestor1 a owl:Class .<br>BIND(REPLACE(str(?ancestor1), '^.*(#|/)', \"\") AS ?ancestorLocalname)<br>BIND(REPLACE(?ancestorLocalname, \"_\", \":\") as ?ancestorId)<br>}",
                "operationId": "getChildrenById",
                "produces": ["text/html", "application/sparql-results+json", "application/ld+json", "application/sparql-results+xml", "text/tab-separated-values", "text/csv", "rdf/xml", "text/turtle"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "The ID of an ontological resource (e.g. GO:0003824)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "number of the page of the result set to display (from 1 and greater than 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_pageSize",
                        "in": "query",
                        "description": "size of a page of the result set (> 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_format",
                        "in": "path",
                        "description": "format in which the results set will be return",
                        "required": true,
                        "type": "string",
                        "enum": [
                            ".json",
                            ".sparql-json",
                            ".xml",
                            ".tsv",
                            ".csv",
                            ".rdf",
                            ".html",
                            ".ttl"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/description{_format}": {
            "post": {
                "tags": [
                    "graphs"
                ],
                "summary": "Retrieve complete description of a resource from AgroLD in JSON format",
                "description": "",
                "operationId": "getdescription",
                "produces": ["text/html", "application/sparql-results+json", "application/ld+json", "application/sparql-results+xml", "text/tab-separated-values", "text/csv", "rdf/xml", "text/turtle"],
                "parameters": [
                    {
                        "name": "uri",
                        "in": "query",
                        "description": "URI of the resource",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "number of the page of the result set to display (from 1 and greater than 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_pageSize",
                        "in": "query",
                        "description": "size of a page of the result set (> 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_format",
                        "in": "path",
                        "description": "format in which the results set will be return",
                        "required": true,
                        "type": "string",
                        "enum": [
                            ".json",
                            ".sparql-json",
                            ".xml",
                            ".tsv",
                            ".csv",
                            ".rdf",
                            ".html",
                            ".ttl"
                        ]
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/graphs{_format}": {
            "post": {
                "tags": [
                    "graphs"
                ],
                "summary": "list all the graphs of AgroLD",
                "description": "",
                "operationId": "listgraphs",
                "produces": ["text/html", "application/sparql-results+json", "application/ld+json", "application/sparql-results+xml", "text/tab-separated-values", "text/csv", "rdf/xml", "text/turtle"],
                "parameters": [
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "number of the page of the result set to display (from 1 and greater than 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_pageSize",
                        "in": "query",
                        "description": "size of a page of the result set (> 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_format",
                        "in": "path",
                        "description": "format in which the results set will be return",
                        "required": true,
                        "type": "string",
                        "enum": [
                            ".json",
                            ".sparql-json",
                            ".xml",
                            ".tsv",
                            ".csv",
                            ".rdf",
                            ".html",
                            ".ttl"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Invalid HTTP method"
                    }
                }
            }
        },
        "/qtls{_format}": {
            "post": {
                "tags": [
                    "qtl"
                ],
                "summary": "Retrieve complete URI and description of all QTLs from AgroLD in JSON format",
                "description": "",
                "operationId": "getQTL",
                "produces": ["text/html", "application/sparql-results+json", "application/ld+json", "application/sparql-results+xml", "text/tab-separated-values", "text/csv", "rdf/xml", "text/turtle"],
                "parameters": [
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "number of the page of the result set to display (from 1 and greater than 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_pageSize",
                        "in": "query",
                        "description": "size of a page of the result set (> 0 or else all results will be displayed)",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "_format",
                        "in": "path",
                        "description": "format in which the results set will be return",
                        "required": true,
                        "type": "string",
                        "enum": [
                            ".json",
                            ".sparql-json",
                            ".xml",
                            ".tsv",
                            ".csv",
                            ".rdf",
                            ".html",
                            ".ttl"
                        ]
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
...
}